* Changed:   * update MsUtility version to 0.0.13   * moved search/filter and search/data to MsUtility version to 0.0.13   * deleted unused application.properties, application-dev.properties, application-com.properties

+++ src/main/java/com/motaharinia/msjpautility/search/SearchTools.java
@@ -41,7 +41,7 @@ public interface SearchTools {
         SearchDataDto<Object[]> searchDataDto = new SearchDataDto();
         searchDataDto.setPageNo(searchFilterDto.getPageNo());
         searchDataDto.setTotalRecordSize(viewPage.getTotalElements());
-        searchDataDto.setTotalPageSize((long) viewPage.getTotalPages());
+        searchDataDto.setTotalPageSize( viewPage.getTotalPages());
         if (userDataMap != null) {
             searchDataDto.setUserDataMap(userDataMap);
         }
